Changed the 'goback' a lot.
Now it remembers the urls inside the app instead of relying on the browser history.

There was a bug where if you navigated to google.com, then to the login screen, then logged in .. it'd just put you back in google.com .. not what we want.

Now if there is no history in the app it'll put you in the home page.

The only disadvantage I can think of is .. if you navigate around the app by typing in to the location bar, you'll get a new App instance every time, so it won't record history in that case.