Before it was compiling each test into a dynamic library (.so) and linking them all into the test runner.
The problem was only the first test suite was getting run.

Now each test suite is compiled into it's own executable.

Also we're integerating into CTest .. so new build instructions are:

    mkdir build
    cd build
    cmake ..
    make -j2
    ctest -j2

or you can use: make test

instead of ctest .. but that doesn't support -j2 .. so ctest -j2 is better :)

To read the test results:

    view Testing/Temporary/LastTest.log